Reformats part of the code
