+++ component-toolbox/src/App.vue
@@ -8,10 +8,10 @@
   </v-app>
 </template>
 
-<script>
+<script lang="ts">
   import {Component, Vue} from "vue-property-decorator";
 
-  @Component()
+  @Component
   export default class App extends Vue {
     mounted() {
       console.log(this.$vuetify.breakpoint.name)
