+++ component-toolbox/src/components/Button.vue
@@ -26,10 +26,10 @@
     </v-dialog>
   </div>
 </template>
-<script>
+<script lang="ts">
   import {Component, Vue} from "vue-property-decorator";
 
-  @Component()
+  @Component
   export default class Button extends Vue {
     dialog = false;
     label = 'Button';
